% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile_model}
\alias{compile_model}
\title{Compiles a stan model}
\usage{
compile_model(model_id, auto_write = TRUE, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
A compiled stan model
}
\description{
Compiles a stan model
}
