% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_model_directory}
\alias{create_model_directory}
\title{Creates a model entry in the library}
\usage{
create_model_directory(model_id, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Creates a model entry in the library
}
