% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_rstan_workflow}
\alias{create_rstan_workflow}
\title{Creates a new model workflow}
\usage{
create_rstan_workflow(model_id, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Creates a new model workflow
}
