% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_ignore_compiled_model}
\alias{git_ignore_compiled_model}
\title{Adds a compiled stan model to project .gitignore}
\usage{
git_ignore_compiled_model(model_id, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Adds a compiled stan model to project .gitignore
}
