% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_object_path}
\alias{model_object_path}
\title{Defines the file path for a model object.}
\usage{
model_object_path(model_id, object, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
The file path for the model
}
\description{
Defines the file path for a model object.
}
