% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{model_path}
\alias{model_path}
\title{Defines the file path for a model in the model library.}
\usage{
model_path(model_id, path = NULL)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
The file path for the model
}
\description{
Defines the file path for a model in the model library.
}
