% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{set_model_library_path}
\alias{set_model_library_path}
\title{Fix a path in global options for stan models to be saved.}
\usage{
set_model_library_path(path)
}
\arguments{
\item{path}{A file path}
}
\value{
Invisible
}
\description{
Fix a path in global options for stan models to be saved.
}
