% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{stan_data_template}
\alias{stan_data_template}
\title{Creates a stan data template file.}
\usage{
stan_data_template(model_id)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Creates a stan data template file.
}
