% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{stan_template}
\alias{stan_template}
\title{Creates a template stan file.}
\usage{
stan_template(model_id)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Creates a template stan file.
}
