% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_model_templates}
\alias{write_model_templates}
\title{Writes standard template files}
\usage{
write_model_templates(model_id, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Writes standard template files
}
