% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_stan_data_template}
\alias{write_stan_data_template}
\title{Writes a template stan data file}
\usage{
write_stan_data_template(model_id, ...)
}
\arguments{
\item{model_id}{A name for a model}
}
\value{
Invisible
}
\description{
Writes a template stan data file
}
